+++ src/Reflection/Entity/PropertyEntity.php
@@ -15,10 +15,8 @@ declare(strict_types=1);
 namespace PhpDocMaker\Reflection\Entity;
 
 use PhpDocMaker\Reflection\AbstractEntity;
-use PhpDocMaker\Reflection\Entity\Traits\DeprecatedTrait;
 use PhpDocMaker\Reflection\Entity\Traits\GetDeclaringClassTrait;
 use PhpDocMaker\Reflection\Entity\Traits\GetTypeAsStringTrait;
-use PhpDocMaker\Reflection\Entity\Traits\SeeTagsTrait;
 use PhpDocMaker\Reflection\Entity\Traits\VisibilityTrait;
 use Roave\BetterReflection\Reflection\ReflectionProperty;
 
@@ -28,10 +26,8 @@ use Roave\BetterReflection\Reflection\ReflectionProperty;
  */
 class PropertyEntity extends AbstractEntity
 {
-    use DeprecatedTrait;
     use GetDeclaringClassTrait;
     use GetTypeAsStringTrait;
-    use SeeTagsTrait;
     use VisibilityTrait;
 
     /**